feat: added beta animations for index page

+++ data/handleImports.ts
@@ -1,31 +1,35 @@
+import { convertFileData } from "@utils/convertFileData";
 import fg from "fast-glob";
-import fs from "fs";
 
-export async function importFiles() {
+function _getPathData(path: string) {
+  const fileNameWithoutExt: string = new RegExp(/(.*)\.import\.txt/g).exec(path)?.[1] as string;
+  const newFileName = `${fileNameWithoutExt}.data.ts`;
+  return {
+    newName: newFileName,
+    withoutExt: fileNameWithoutExt,
+  };
+}
+
+export async function importNewFiles() {
+  if (window) return;
+  const fs = (await import("fs")).default;
   const found = await fg("**/*.import.txt", { objectMode: true });
   for (const file of found) {
     const fileData = fs.readFileSync(file.path, { encoding: "utf-8" });
-    const fileLines = fileData.split("\n");
-    const fileNameWithoutExt: string = new RegExp(/(.*)\.import\.txt/g).exec(file.name)?.[1] as string;
-    const newFileName = `${fileNameWithoutExt}.data.ts`;
-    const data = [];
-    for (const line of fileLines) {
-      const [answer, question] = line.split(" - ");
-      data.push({ question, answer });
-    }
-    const dataAsString = JSON.stringify(data, null, 2);
+    const stringData = await convertFileData(fileData);
+    const { newName, withoutExt } = _getPathData(file.path);
     fs.writeFileSync(
-      `data/${newFileName}`,
-      `const wordsData = ${dataAsString}
+      `data/${newName}`,
+      `const wordsData = ${stringData}
 
-const exported = {
-  id: "${fileNameWithoutExt}",
-  data: wordsData,
-  lang: "lang",
-  tags: ["tag"],
-};
+      const exported = {
+        id: "${withoutExt}",
+        data: wordsData,
+        lang: "lang",
+        tags: ["tag"],
+      };
 
-export default exported;`
+      export default exported;`
     );
     fs.unlinkSync(file.path);
   }
